vim: Move some plugins from pack/*/opt to pack/*/start
YouCompleteMe is a badly behaved plugin which should only be loaded as
needed, so I'll keep relying on packadd for that one.

The other plugins I moved at most set a few variables and define some
commands.
vim-lsp will start disabled if g:lsp_auto_enable is 0. On the language
modes I want, I'll enable it.